icalendar: only run the changelog test.
See https://github.com/collective/icalendar/pull/514#issuecomment-1506108571 and following comments.
From a Plone core development perspective, `icalendar` is a third party package since a while.
It is used a lot outside of Plone, and is taken care of by people who are not core Plone developers.

The `plone.app.event` tests are run on every PR of this package, see [yaml file](https://github.com/collective/icalendar/blob/v5.0.5/.github/workflows/tests.yml#L26) and [`tox.ini`](https://github.com/collective/icalendar/blob/v5.0.5/tox.ini#L27).
I keep tabs on new versions, and manually update its version in `buildout.coredev`.
So nothing is needed on Jenkins, so `mr.roboto` does not need to do anything.  I have removed the roboto hooks in the settings of the `icalendar` repo.

But: checking for a changelog entry is still wanted.  That is where this PR comes in:

1. Add `icalendar` to all ignore lists, except for the required changelog.  (Also sort them alphabetically.)
2. Fix the `githubcommithooks` endpoint to look for a single repo in the collective if it is not in Plone.

With this in place, I can enable the hooks again for this single repo, and only the changelog entry would be checked.