cleanup(hooks): only for buildout.coredev
On each GitHub repository we have a few web hooks that call mr.roboto,
but one of them is talking directly to jenkins.

That one is actually only meant for buildout.coredev, to trigger the
main jenkins jobs on any change there. We don't want that to happen for
any other repository, otherwise the queue would grow out of bounds.

This PR ensures that whenever web hooks are re-created on GitHub, only
the web hooks meant for each repository are added, and skips the one
meant for buildout.coredev, which is added only for buildout.coredev.